feat: ✨ spinner
- 1.5초 지연 강제로 spinner 보이게함
