+++ frontend/src/collections/HomeCollection/index.tsx
@@ -6,6 +6,7 @@ import RootStore from '@/stores/RootStore';
 import Search from '@/components/Search';
 import TimeLine from '@/components/TimeLine';
 import BalanceGame from '@/components/BalanceGame';
+import Spinner from '@/components/Spinner';
 
 interface IHomeCollectionProps {}
 
@@ -14,7 +15,7 @@ const HomeCollection = () => {
   const { homeStore: store, uiStore, searchResultStore } = rootStore;
   const { timeLine } = store;
 
-  const content = !rootStore.isLoading && timeLine.days.length > 0 ? <TimeLine timeLine={timeLine} /> : null;
+  const content = !rootStore.isLoading && timeLine.days.length > 0 ? <TimeLine timeLine={timeLine} /> : <Spinner />;
 
   return (
     <>
