+++ frontend/src/css/components/z-index.scss
@@ -1,4 +1,3 @@
-$VOTE_SPINNER_Z_INDEX: 2;
 $SEARCH_BOX_Z_INDEX: 2;
 $HEADER_Z_INDEX: 9;
 $HEADER_TITLE_Z_INDEX: 9;
