--- frontend/src/css/spinner/vote-spinner.scss
@@ -1,6 +1,0 @@
-.vote-spinner {
-  position: fixed;
-  left: 50%;
-  top: 30%;
-  transform: translate(-50%, 0);
-}
