+++ frontend/src/repositories/HomeRepository.ts
@@ -1,4 +1,6 @@
 import API from '@/utils/api';
+import { SPINNER_API_WAITING_TIME } from '@/utils/constants';
+import { wait } from '@/utils/utils';
 
 class HomeRepository {
   URL = process.env.ENV === 'local' ? '/articles' : '/v1/articles';
@@ -7,9 +9,9 @@ class HomeRepository {
     this.URL = url || this.URL;
   }
 
-  getArticle() {
+  async getArticle() {
     const currentTime = new Date().getTime();
-    return API.get(`${this.URL}?_=${currentTime}`);
+    return Promise.all([API.get(`${this.URL}?_=${currentTime}`), wait(SPINNER_API_WAITING_TIME)]);
   }
 }
