+++ frontend/src/utils/constants.ts
@@ -1,0 +1,5 @@
+const SPINNER_API_WAITING_TIME = 1500;
+
+const TEST_DEFAULT_TEMPLATE = '';
+
+export { SPINNER_API_WAITING_TIME, TEST_DEFAULT_TEMPLATE };
